requirements.txt
telethon>=1.24.0
python-telegram-bot>=20.3
asyncio-mqtt>=0.13.0
python-dotenv>=1.0.0
pystray>=0.19.4
Pillow>=9.5.0